sidescroll and menus implementation

+++ src/main/java/entities/Entity.java
@@ -23,10 +23,10 @@ public abstract class Entity {
 
   protected void drawHitbox(Graphics g, int lvlOffset) {
     g.setColor(Color.PINK);
-    g.drawRect((int) hitbox.x- lvlOffset, (int) hitbox.y, (int) hitbox.width, (int) hitbox.height);
+    g.drawRect((int) hitbox.x - lvlOffset, (int) hitbox.y, (int) hitbox.width, (int) hitbox.height);
   }
 
-  protected void initHitbox(float x, float y, float width, float height) {
+  protected void initHitbox(float x, float y, int width, int height) {
     hitbox = new Rectangle2D.Float(x, y, height, width);
   }
